requirements.txt
numpy==1.23.1
pandas==1.4.3
scikit-surprise==1.1.1
scikit-learn==1.1.1
PyYAML==6.0
argparse==1.4.0